requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.13.0
opencv-python==4.8.0.74
RPi.GPIO==0.7.1
picamera==1.13
Adafruit_DHT==1.4.0
joblib==1.3.1
matplotlib==3.7.2
